backend/requirements.txt
fastapi==0.115.1
uvicorn==0.34.0
prisma==0.15.0
python-dotenv==1.1.0
httpx==0.28.1
pydantic==2.11.1
openai==1.30.3
stripe==9.9.0
firebase-admin==6.5.0

# 開発用
black==23.12.0
pylint==3.0.3
